Fixed media streaming - use HTTP_RANGE headers
This should fix problems with various players, for longer songs/videos,
while playing and trying to do a seek.
In that moment, the player will use the play url but with HTTP_RANGE http
headers, which current implementation wasn't using.
Inspired from code of Ampache at ampache.org.
Tested with amarok 2.6-git on a ~2hours "song".