Use img instead of css to display background image
With css, the background image is "cut" by other div (for example, you
may not see the background image in the files list if there is too many
file rows).
I use an img, with position absolute, bottom right like the css
background and small opacity to prevent visibility problems.