Update contacts.css
Proposed fix for #602 (Sharing contacts)
--------------------------
This ist a partial fix. In german locale the translated text "Dateien für den Import auswählen" is still too long so the upload button ist partially hidden by the new scrollbar. That should be fixed in the affected locale(s).

I'm new to git(hub) an so i hope this is the right place for this. Sorry for any inconveniences...