fixing aspect-ratio-bug for ie10 in slideshow.js
Setting both height and width using px-Values instead of just Setting one of the attributes to "100%". Without this fix IE10 does not keep the aspect-ratio of the image. 

Just registered to github to give this feedback. Please re-test. Test coverage with ie10: 50% of the new code. Not tried with other browsers.