improved navigation
When the Image finished loading it was displayed by the callback of showImage, even if the current Image was changed in the meantime. This bug should corrected by this commit.

The code for handling image-load-Errors was changed. When loading fails the deferred-Object fails. The fail method could be used to set a callback to display an error. Currently nothing is changed in the UI. After load has failed the spinner is displayed until the next image is loaded.