Fixed support for IdP initiated logout
Works by saving whether a user has been logged in using saml in a
cookie and later on checking if this cookie is set but simplesamlphp
reports the user as unauthenticated. If this is the case trigger a
logout. Due to a design limitation in owncloud this check is performed
in the constructor of the saml user backend which is called for every
request.