Added ODF viewer mime type exclusion when documents app is active
Whenever the ODF viewer and documents app are active at the same time,
the ODF viewer leaves the file action free for the documents app.

This way clicking files with the mime types supported by the documents app
will open the documents app instead of the ODF viewer.

This is for https://github.com/owncloud/documents/issues/71