Implement request #301: Create users when using external auth login
After a successful login with imap, smb or ftp, the corresponding
user is created in an own database table.
This is needed to make e.g. sharing files with this users work.

The table may be used by several different external user authentication
services.

See the discussion in https://github.com/owncloud/apps/pull/1579
for details on the implementation design.